EdgeWeights->doubles

+++ include/graph.h
@@ -19,16 +19,16 @@ class Graph {
 		Graph();
 		Graph(std::string filename);
 		Graph(int numNodes, int numEdges);
-		double getEdgePoint(int i,int j);
+		int getEdgePoint(int i,int j);
 		int getNumNodes();
 		int getNumEdges();
                 int getNodeWeight(int node);
-                int getEdgeWeight(int node, int neighborInd);
+                double getEdgeWeight(int node, int neighborInd);
                 int getNodeMatch(int node);
                 void setNodeMatch(int node, int val);
                 std::vector<int> getMatchList();
                 std::vector<int> getNeighbors(int node);
-                std::vector<int> getEdges(int node);
+                std::vector<double> getEdges(int node);
 		std::vector<int> reorderGraph(std::vector<int> indMap);
                 std::vector<int> getChildren(int node);
 		int coarsenFrom(Graph & g);
@@ -40,7 +40,7 @@ class Graph {
 		int numEdges, numNodes, numChildren;
 		std::vector<std::vector<int> > edge;
 		std::vector<std::vector<int> > neighborList;
-                std::vector<std::vector<int> > edgeWeights;
+                std::vector<std::vector<double> > edgeWeights;
                 std::vector<int> nodeWeights;
                 std::vector<int> matchList;
                 std::vector<std::vector<int> > parentList;
